// src/templates/templates.enum.ts
export enum Templates {
  VERIFY_EMAIL_OWNER = 'verify-email-owner',
  VERIFY_EMAIL_CONTRACTOR = 'verify-email-contractor',
  DRIVER_CREDS = 'driver-creds',
  FORGOT_PASSWORD = 'forgot-password',
  SEND_EMAIL_WITH_PASSWORD = 'send-email-with-password',
  NEW_JOB_DRIVER = 'new-job-driver',
  NEW_DRIVER_ASSIGNED = 'new-driver-assigned',
  JOB_COMPLETE = 'job-complete',
  DRIVER_STARTED_JOB = 'driver-started-job',
  NEW_JOB_IN_THE_AREA = 'new-job-in-the-area',
  PROBLEM_REPORTED = 'problem-reported',
  CANCELED_JOB = 'canceled-job',
  NEW_DISPUTE = 'new-dispute',
  ADDED_INTERESTS = 'added-interests',
  SOLVED_DISPUTE = 'solved-dispute',
  FINISHED_JOB_CONTRACTOR = 'finished-job-contractor',
  DISABLE_OWNER = 'disable-owner',
  DISABLE_CONTRACTOR = 'disable-contractor',
  CONTRACTOR_ACCOUNT_VERIFIFED_BY_ADMIN = 'contractor-account-verified-by-admin',
  OWNER_ACCOUNT_VERIFIFED_BY_ADMIN = 'owner-account-verified-by-admin',
  DISPATCHER_CREDS = 'dispatcher-creds',
  DRIVER_WEEKLY_SUMMARY = 'driver-weekly-summary',
  FINISHED_JOB_OWNER = 'finished-job-owner',
  SCHEDULED_JOB_FINISHED_OWNER = 'scheduled-job-finished-owner',
  SCHEDULED_JOB_FINISHED_CONTRACTOR = 'scheduled-job-finished-contractor',
  ASSIGNATION_FINISHED_EMAIL = 'assignation-finished-email',
  PAYMENT_RECEIVED = 'payment-received',
  NEW_DISPUTE_OWNER = 'new-dispute-owner',
  NEW_DISPUTE_CONTRACTOR = 'new-dispute-contractor',
  DISPUTE_RESOLVED = 'dispute-resolved',
  JOB_NOT_STARTED = 'job-not-started',
  JOB_NOT_FILLED_OUT = 'job-not-filled-out',
  FUTURE_SURCHARGE = 'future-surcharge',
  CASH_ADVANCE_REVIEW = 'cash-advance-review',
  NEW_TRUCK_REVIEW = 'new-truck-review',
  AUTOMATICALLY_CLOKED_OUT = 'automatically-clocked-out',
  AUTOMATICALLY_CLOKED_OUT_OWNER = 'automatically-clocked-out-owner',
  ASSIGNATION_FINISHED_AUTOMATICALLY = 'assignation-finished-automatically',
  REQUEST_TRUCK = 'request-truck',
  NEW_PASSWORD = 'new-password',
}